Bring issue logic into the fold
